move functions from hypergraph to hon
